move timer from ctx to separate file #68
