+++ lib/app/timer.ex
@@ -1,0 +1,117 @@
+defmodule App.Timer do
+  use Ecto.Schema
+  import Ecto.Changeset
+  import Ecto.Query, warn: false
+  alias App.Repo
+  alias __MODULE__
+
+  schema "timers" do
+    field :end, :naive_datetime
+    field :start, :naive_datetime
+    field :item_id, :id
+    field :person_id, :id
+
+    timestamps()
+  end
+
+  @doc false
+  def changeset(timer, attrs) do
+    timer
+    |> cast(attrs, [:start, :end])
+    |> validate_required([:start, :end])
+  end
+
+  @doc """
+  Returns the list of timers.
+
+  ## Examples
+
+      iex> list_timers()
+      [%Timer{}, ...]
+
+  """
+  def list_timers do
+    Repo.all(Timer)
+  end
+
+  @doc """
+  Gets a single timer.
+
+  Raises `Ecto.NoResultsError` if the Timer does not exist.
+
+  ## Examples
+
+      iex> get_timer!(123)
+      %Timer{}
+
+      iex> get_timer!(456)
+      ** (Ecto.NoResultsError)
+
+  """
+  def get_timer!(id), do: Repo.get!(Timer, id)
+
+  @doc """
+  Creates a timer.
+
+  ## Examples
+
+      iex> create_timer(%{field: value})
+      {:ok, %Timer{}}
+
+      iex> create_timer(%{field: bad_value})
+      {:error, %Ecto.Changeset{}}
+
+  """
+  def create_timer(attrs \\ %{}) do
+    %Timer{}
+    |> Timer.changeset(attrs)
+    |> Repo.insert()
+  end
+
+  @doc """
+  Updates a timer.
+
+  ## Examples
+
+      iex> update_timer(timer, %{field: new_value})
+      {:ok, %Timer{}}
+
+      iex> update_timer(timer, %{field: bad_value})
+      {:error, %Ecto.Changeset{}}
+
+  """
+  def update_timer(%Timer{} = timer, attrs) do
+    timer
+    |> Timer.changeset(attrs)
+    |> Repo.update()
+  end
+
+  @doc """
+  Deletes a Timer.
+
+  ## Examples
+
+      iex> delete_timer(timer)
+      {:ok, %Timer{}}
+
+      iex> delete_timer(timer)
+      {:error, %Ecto.Changeset{}}
+
+  """
+  def delete_timer(%Timer{} = timer) do
+    Repo.delete(timer)
+  end
+
+  @doc """
+  Returns an `%Ecto.Changeset{}` for tracking timer changes.
+
+  ## Examples
+
+      iex> change_timer(timer)
+      %Ecto.Changeset{source: %Timer{}}
+
+  """
+  def change_timer(%Timer{} = timer) do
+    Timer.changeset(timer, %{})
+  end
+end
