+++ test/app/ctx_test.exs
@@ -126,71 +126,4 @@ defmodule App.CtxTest do
       assert %Ecto.Changeset{} = Ctx.change_status(status)
     end
   end
-
-  describe "timers" do
-    alias App.Ctx.Timer
-
-    @valid_attrs %{end: ~N[2010-04-17 14:00:00], start: ~N[2010-04-17 14:00:00]}
-    @update_attrs %{
-      end: ~N[2011-05-18 15:01:01],
-      start: ~N[2011-05-18 15:01:01]
-    }
-    @invalid_attrs %{end: nil, start: nil}
-
-    def timer_fixture(attrs \\ %{}) do
-      {:ok, timer} =
-        attrs
-        |> Enum.into(@valid_attrs)
-        |> Ctx.create_timer()
-
-      timer
-    end
-
-    test "list_timers/0 returns all timers" do
-      timer = timer_fixture()
-      assert Ctx.list_timers() == [timer]
-    end
-
-    test "get_timer!/1 returns the timer with given id" do
-      timer = timer_fixture()
-      assert Ctx.get_timer!(timer.id) == timer
-    end
-
-    test "create_timer/1 with valid data creates a timer" do
-      assert {:ok, %Timer{} = timer} = Ctx.create_timer(@valid_attrs)
-      assert timer.end == ~N[2010-04-17 14:00:00]
-      assert timer.start == ~N[2010-04-17 14:00:00]
-    end
-
-    test "create_timer/1 with invalid data returns error changeset" do
-      assert {:error, %Ecto.Changeset{}} = Ctx.create_timer(@invalid_attrs)
-    end
-
-    test "update_timer/2 with valid data updates the timer" do
-      timer = timer_fixture()
-      assert {:ok, %Timer{} = timer} = Ctx.update_timer(timer, @update_attrs)
-      assert timer.end == ~N[2011-05-18 15:01:01]
-      assert timer.start == ~N[2011-05-18 15:01:01]
-    end
-
-    test "update_timer/2 with invalid data returns error changeset" do
-      timer = timer_fixture()
-
-      assert {:error, %Ecto.Changeset{}} =
-               Ctx.update_timer(timer, @invalid_attrs)
-
-      assert timer == Ctx.get_timer!(timer.id)
-    end
-
-    test "delete_timer/1 deletes the timer" do
-      timer = timer_fixture()
-      assert {:ok, %Timer{}} = Ctx.delete_timer(timer)
-      assert_raise Ecto.NoResultsError, fn -> Ctx.get_timer!(timer.id) end
-    end
-
-    test "change_timer/1 returns a timer changeset" do
-      timer = timer_fixture()
-      assert %Ecto.Changeset{} = Ctx.change_timer(timer)
-    end
-  end
 end
